site: refactor demo actions code (#54033)
